Final refactor of extensions.
